task 7

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -3,7 +3,7 @@
 type-annotated function sum_mixed_list
 """
 from typing import List, Union
-s = Union[float, int]
+s = Union[int, float]
 
 
 def sum_mixed_list(mxd_lst: List[s]) -> float:
